relocate button component
