+++ src/ts/components/buttons/type.ts
